fix: 테스트 라우터, 컨트롤러 함수 파일 수정
removeAllMembersFromGroup 함수에서 그룹에 대기 중인 모든 사용자 Id 를 가져와서 대기 그룹목록에서도 삭제
