+++ server/controller/ctrGroup.js
@@ -1,78 +1,41 @@
-const { Group, User, GroupMember } = require('../models');
-
-const Sequelize = require('sequelize');
+const { User, Group } = require('../schemas/schema');
+const mongoose = require('mongoose');
 
 // 카테고리에 따른 그룹 목록을 반환하는 함수
 exports.getCategoryGroups = async (req, res) => {
   try {
-    // 태균
-    // 토큰에서 현재유저 user_id 가져오기
-    const userId = res.locals.decoded.userInfo.id;
+    // 토큰에서 현재 유저 정보 가져오기
+    const userInfo = res.locals.decoded.userInfo;
+
+    if (!userInfo) {
+      return res.status(400).send({
+        isSuccess: false,
+        code: 400,
+        error: '사용자 정보를 찾을 수 없습니다.',
+      });
+    }
+
+    const userId = userInfo.id;
+    const userCategory = userInfo.category;
 
     // 사용자의 카테고리를 조회
-    const userCategory = await User.findOne({
-      where: { user_id: userId },
-      attributes: ['user_category_name'],
-    });
-
-    // 사용자 카테고리 데이터가 존재하는 경우
-    if (userCategory) {
-      const userCategoryName = userCategory.dataValues.user_category_name;
-      console.log('카테고리는', userCategoryName);
-
-      // 사용자의 카테고리와 일치하는 그룹을 조회
-      const group = await Group.findAll({
-        where: { group_category_name: userCategoryName },
-      });
-
-      let selectedCategoryGroup;
-
-      // 일치하는 그룹이 하나 이상 있는 경우
-      if (group.length > 0) {
-        selectedCategoryGroup = group.map(group => group.dataValues);
-        // 선택한 카테고리에 속하는 그룹이 하나 이상 있는 경우
-        if (selectedCategoryGroup.length > 0) {
-          const groupIds = selectedCategoryGroup.map(group => group.group_id);
-
-          const groupManager = await GroupMember.findOne({
-            where: { group_id: groupIds, is_admin: true },
-            attributes: ['user_id'],
-          });
-
-          if (groupManager) {
-            // 그룹 관리자가 존재할 때
-            res.status(200).send({
-              isSuccess: true,
-              code: 200,
-              study_groups: selectedCategoryGroup,
-              groupManager: groupManager.dataValues.user_id,
-            });
-          } else {
-            // 그룹관리자가 존재하지 않을 때
-            res.status(404).send({
-              isSuccess: false,
-              code: 404,
-              error: '그룹 관리자를 찾을 수 없습니다.',
-            });
-          }
-        } else {
-          // 해당 카테고리의 스터디 그룹이 존재하지 않을 때
-          res.status(204).send({
-            isSuccess: false,
-            code: 204,
-            error: '해당하는 카테고리의 그룹이 없습니다.',
-          });
-        }
-      } else {
-        // 해당 카테고리의 스터디 그룹이 존재하지 않을 때
-        selectedCategoryGroup = '해당하는 카테고리의 그룹이 없습니다.';
-        res.status(204).send({
-          isSuccess: false,
-          code: 204,
-          study_groups: selectedCategoryGroup,
-          error: '해당하는 카테고리의 그룹이 없습니다.',
+    const user = await User.findById(userId);
+    if (user) {
+      const groups = await Group.find({ group_category: userCategory });
+      if (groups.length > 0) {
+        res.status(200).send({
+          isSuccess: true,
+          code: 200,
+          study_groups: groups,
         });
       }
+      console.log(groups);
+    } else {
+      res.status(204).send({
+        isSuccess: false,
+        code: 204,
+        error: '해당하는 카테고리의 그룹이 없습니다.',
+      });
     }
   } catch (err) {
     // 에러가 발생한 경우 서버 오류 메시지와 HTTP 상태 코드 500 반환
@@ -84,139 +47,242 @@ exports.getCategoryGroups = async (req, res) => {
 // 사용자별 그룹 목록을 반환하는 함수
 exports.getCategoryGroupsByUser = async (req, res) => {
   try {
-    // 태균
-    // 토큰에서 현재유저 user_id 가져오기
-    const userId = res.locals.decoded.userInfo.id;
+    // 토큰에서 현재 유저 정보 가져오기
+    const userInfo = res.locals.decoded.userInfo;
 
-    // 사용자의 그룹 ID 목록을 조회
-    const userGroupIds = await GroupMember.findAll({
-      where: { user_id: userId },
-      attributes: ['group_id'],
-    });
-
-    const userGroupsData = [];
-
-    // 사용자가 속한 그룹이 하나 이상 있는 경우
-    if (userGroupIds.length > 0) {
-      const groupIds = userGroupIds.map(member => member.dataValues.group_id);
-
-      // 그룹 ID 목록을 사용하여 사용자의 그룹 목록을 조회
-      const userGroups = await Group.findAll({
-        where: { group_id: groupIds },
+    if (!userInfo) {
+      return res.status(400).send({
+        isSuccess: false,
+        code: 400,
+        error: '사용자 정보를 찾을 수 없습니다.',
       });
-
-      // 조회된 그룹 데이터를 사용자가 속한 그룹 데이터로 변환
-      userGroupsData.push(...userGroups.map(group => group.dataValues));
-
-      // 그룹 관리자를 찾아 배열로 초기화
-      const groupManager = await GroupMember.findAll({
-        where: { group_id: groupIds, is_admin: true },
-        attributes: ['user_id'],
-      });
-
-      // 사용자가 속한 그룹 관리자 목록을 추가
-      userGroupsData.push({ groupManager: groupManager.map(manager => manager.dataValues.user_id) });
     }
 
-    // HTTP 상태 코드 200과 사용자 그룹 데이터를 JSON 응답으로 반환
-    res.status(200).send({ isSuccess: true, code: 200, study_groups: userGroupsData });
+    const userId = userInfo.id;
+
+    // 사용자의 그룹 ID 목록 조회
+    const userGroup = await User.findById(userId).select('groups');
+    let groups = []; // 변수 선언 및 초기화
+    console.log('userGroup은 ', userGroup);
+    if (userGroup) {
+      console.log('사용자의 그룹 ID 목록:', userGroup.groups);
+
+      // 그룹 ID 목록을 사용하여 그룹 정보 조회 (비동기 처리)
+      groups = await Group.find({ _id: { $in: userGroup.groups } });
+      console.log('조회된 그룹 정보:', groups);
+      // 여기서 groups에 조회된 그룹 정보가 배열로 포함됩니다.
+    } else {
+      console.log('사용자를 찾을 수 없습니다.');
+      // 사용자를 찾지 못한 경우에 대한 처리
+    }
+    res.status(200).send({ isSuccess: true, code: 200, study_groups: groups });
   } catch (err) {
     console.error(err);
     // 에러가 발생한 경우 서버 오류 메시지와 HTTP 상태 코드 500 반환
     res.status(500).send({ isSuccess: false, code: 500, error: '서버에서 오류가 발생했습니다.' });
   }
 };
+
 //그룹을 요청하는 함수
 exports.joinGroupRequest = async (req, res) => {
-  const session = await mongoose.startSession();
-  session.startTransaction();
-
   try {
-    const userId = res.locals.decoded.userInfo.id; // 유저 objectId
-    const { groupId } = req.params;
+    // 토큰에서 현재 유저 정보 가져오기
+    const userInfo = res.locals.decoded.userInfo;
 
-    const updatedUserData = {
-      pending_groups: [
-        {
-          group: groupId, // groupId는 해당 그룹의 ObjectId
-        },
-      ],
-    };
+    if (!userInfo) {
+      return res.status(400).send({
+        isSuccess: false,
+        code: 400,
+        error: '사용자 정보를 찾을 수 없습니다.',
+      });
+    }
 
-    const updatedGroupData = {
-      join_requests: [
-        {
-          user_id: userId, // userId는 그룹을 요청한 유저의 objectId
-        },
-      ],
-    };
+    const userId = userInfo.id;
 
-    // 사용자 및 그룹 업데이트
-    const updatedUser = await User.findByIdAndUpdate(
-      userId,
-      { $push: { pending_groups: updatedUserData } },
-      { new: true }
-    ).session(session);
-    const updatedGroup = await Group.findByIdAndUpdate(
-      groupId,
-      { $push: { join_requests: updatedGroupData } },
-      { new: true }
-    ).session(session);
+    const { groupId } = req.params; // 가입하려는 group의 object Id
 
-    await session.commitTransaction();
-    session.endSession();
+    // 사용자 업데이트
+    const user = await User.findById(userId);
+    if (user) {
+      user.pending_groups.push({ group: groupId });
+      await user.save();
+    }
 
-    console.log('사용자가 업데이트되었습니다.');
-    console.log(updatedUser);
-    console.log('그룹이 업데이트되었습니다.');
-    console.log(updatedGroup);
+    // 그룹 업데이트
+    const group = await Group.findById(groupId);
+    if (group) {
+      group.join_requests.push({ user_id: userId });
+      await group.save();
+    }
+    res.status(200).send({
+      isSuccess: true,
+      message: '그룹 가입 요청이 성공적으로 처리되었습니다.',
+    });
   } catch (err) {
-    await session.abortTransaction();
-    session.endSession();
-
     console.error(err);
-    // 에러 응답 보내거나 다른 작업 수행
+
+    res.status(500).send({
+      isSuccess: false,
+      error: '서버 오류가 발생했습니다.',
+    });
   }
 };
 
+//그룹 요청 수락 기능 함수
+exports.acceptGroupMembershipRequest = async (req, res) => {
+  // 토큰에서 현재 유저 정보 가져오기
+  const userInfo = res.locals.decoded.userInfo;
+
+  if (!userInfo) {
+    return res.status(400).send({
+      isSuccess: false,
+      code: 400,
+      error: '사용자 정보를 찾을 수 없습니다.',
+    });
+  }
+
+  const userId = userInfo.id;
+  const { groupId } = req.params;
+
+  try {
+    const group = await Group.findById(groupId);
+    const user = await User.findById(userId);
+
+    if (!group || !user) {
+      return res.status(404).send({
+        isSuccess: false,
+        error: '그룹 또는 사용자를 찾을 수 없습니다.',
+      });
+    }
+
+    // 그룹의 join_requests에서 requestId에 해당하는 문서를 찾아서 삭제
+    group.join_requests = group.join_requests.filter(request => request.user_id.toString() !== userId);
+
+    // 그룹의 멤버로 사용자 추가
+    group.members.push(userId);
+
+    // 사용자의 pending_groups에서 groupId에 해당하는 문서를 찾아서 삭제
+    user.pending_groups = user.pending_groups.filter(groupRequest => groupRequest.group.toString() !== groupId);
+
+    // 사용자의 그룹 목록에 groupId 추가
+    user.groups.push(groupId);
+
+    await group.save();
+    await user.save();
+
+    return res.status(200).send({
+      isSuccess: true,
+      message: '그룹 멤버십 요청을 성공적으로 수락했습니다.',
+    });
+  } catch (err) {
+    console.error(err);
+
+    return res.status(500).send({
+      isSuccess: false,
+      error: '서버 오류가 발생했습니다.',
+    });
+  }
+};
+// 그룹 요청 거절 기능 함수
+exports.rejectGroupMembershipRequest = async (req, res) => {
+  // 토큰에서 현재 유저 정보 가져오기
+  const userInfo = res.locals.decoded.userInfo;
+
+  if (!userInfo) {
+    return res.status(400).send({
+      isSuccess: false,
+      code: 400,
+      error: '사용자 정보를 찾을 수 없습니다.',
+    });
+  }
+
+  const userId = userInfo.id;
+  const { groupId } = req.params;
+
+  try {
+    const group = await Group.findById(groupId);
+    const user = await User.findById(userId);
+
+    if (!group || !user) {
+      return res.status(404).send({
+        isSuccess: false,
+        error: '그룹 또는 사용자를 찾을 수 없습니다.',
+      });
+    }
+
+    // 그룹의 join_requests에서 requestId에 해당하는 문서를 찾아서 삭제
+    group.join_requests = group.join_requests.filter(request => request.user_id.toString() !== userId);
+
+    // 사용자의 pending_groups에서 groupId에 해당하는 문서를 찾아서 삭제
+    user.pending_groups = user.pending_groups.filter(groupRequest => groupRequest.group.toString() !== groupId);
+
+    await group.save();
+    await user.save();
+
+    return res.status(200).send({
+      isSuccess: true,
+      message: '그룹 멤버십 요청을 성공적으로 거절했습니다.',
+    });
+  } catch (err) {
+    console.error(err);
+
+    return res.status(500).send({
+      isSuccess: false,
+      error: '서버 오류가 발생했습니다.',
+    });
+  }
+};
 // 새 그룹 생성하는 함수
 exports.postGroupInformation = async (req, res) => {
   try {
+    //그룹을 생성하는 유저의 아이디 -> 그룹장
+    // 토큰에서 현재 유저 정보 가져오기
+    const userInfo = res.locals.decoded.userInfo;
+
+    if (!userInfo) {
+      return res.status(400).send({
+        isSuccess: false,
+        code: 400,
+        error: '사용자 정보를 찾을 수 없습니다.',
+      });
+    }
+
+    const userId = userInfo.id;
+
     // 클라이언트에서 요청으로 받은 데이터 추출
-    const { name, password, description, category, dailyGoalTime, maximumNumberMember, isCameraOn } = req.body;
-    // TODO: 그룹장 정보 넣기 (그룹장 정보를 어떻게 처리할지에 대한 내용을 추가로 구현해야 함)
-    // TODO: 유저의 카테고리 그룹생성시 default로 박기
+    const { name, description, category, dailyGoalTime, maximumNumberMember, isCameraOn } = req.body;
+    // TODO: 유저의 카테고리 그룹생성시 default로 박기??
     // TODO: multer file path -> client와 붙이면서 확인
-    // 서버에 저장된 파일 이름
     const { filename } = req.file;
     // path == 이미지를 받을 수 있는 URL
     const imagePath = `/api/static/profileImg/${filename}`;
-
-    // Sequelize 모델을 사용하여 새 그룹 생성
-    const newGroup = await Group.create({
-      name: name, // 그룹 이름
-      password: password, // 비밀번호
-      description: description, // 그룹 설명
-      group_category_name: category, //카테고리
+    const newGroup = new Group({
+      group_leader: userId, //그룹장의 user objectId
+      group_name: name, // 그룹 이름
+      group_description: description, // 그룹 설명
+      group_category: category, //카테고리
       group_image_path: imagePath, //그룹 프로필 이미지
       daily_goal_time: dailyGoalTime, // 일일 목표 시간
-      maximum_number_member: maximumNumberMember, // 최대 회원 수
+      group_maximum_member: maximumNumberMember, // 최대 회원 수
       is_camera_on: isCameraOn, // 카메라 상태
+      members: userId,
     });
 
-    const groupManager = await GroupMember.create({
-      group_id: newGroup.dataValues.group_id,
-      //TODO user_id 토큰에서 가져오기
-      user_id: 1,
-      is_admin: true,
-    });
+    // 그룹 저장
+    const savedGroup = await newGroup.save();
+
+    // 사용자를 찾아서 사용자의 groups 배열에 새 그룹 ObjectId를 추가
+    const user = await User.findById(userId);
+    user.groups.push(savedGroup._id); // 새 그룹의 ObjectId를 추가
+    // 사용자 업데이트 및 저장
+    await user.save();
 
     // HTTP 상태 코드 201 (Created)와 함께 새 그룹 정보를 클라이언트에 반환
     res.status(201).send({
       isSuccess: true,
       code: 201,
       message: '스터디 그룹이 성공적으로 생성되었습니다.',
-      groupManager: groupManager.dataValues.user_id, //user_id값으로 들어감
+      groupLeader: user.nick_name,
       data: newGroup,
     });
   } catch (err) {
@@ -227,13 +293,24 @@ exports.postGroupInformation = async (req, res) => {
 
 exports.patchGroupInformation = async (req, res) => {
   try {
+    // 토큰에서 현재 유저 정보 가져오기
+    const userInfo = res.locals.decoded.userInfo;
+
+    if (!userInfo) {
+      return res.status(400).send({
+        isSuccess: false,
+        code: 400,
+        error: '사용자 정보를 찾을 수 없습니다.',
+      });
+    }
+
+    const userId = userInfo.id;
     // 요청 파라미터에서 그룹 ID를 가져옴
     const { groupId } = req.params;
 
     // 요청 본문에서 그룹 정보 업데이트를 위한 필드들을 가져옴
     const {
       name, // 그룹 이름
-      password, // 비밀번호
       description, // 그룹 설명
       category, // 그룹의 카테고리 이름 (User FK 값)
       imagePath, // 그룹 프로필 이미지 경로
@@ -243,29 +320,24 @@ exports.patchGroupInformation = async (req, res) => {
     } = req.body;
 
     // 그룹 정보 업데이트 수행
-    const patchGroup = await Group.update(
-      {
-        name: name,
-        password: password,
-        description: description,
-        group_category_name: category,
-        group_image_path: imagePath,
-        daily_goal_time: dailyGoalTime,
-        maximum_number_member: maximumNumberMember,
-        is_camera_on: isCameraOn,
-      },
-      { where: { group_id: groupId } }
-    );
-
-    // 업데이트된 그룹 정보 조회
-    const modifiedGroup = await Group.findOne({
-      where: { group_id: groupId },
-    });
-
-    if (patchGroup) {
+    const updatedFields = {
+      group_name: name,
+      group_description: description,
+      group_category: category,
+      group_image_path: imagePath,
+      daily_goal_time: dailyGoalTime,
+      group_maximum_member: maximumNumberMember,
+      is_camera_on: isCameraOn,
+    };
+    //그룹 업데이트
+    const updatedGroup = await Group.findByIdAndUpdate(groupId, updatedFields, { new: true });
+    if (updatedGroup) {
       // 업데이트가 성공했을 경우
-      res.status(200).send({ isSuccess: true, code: 200, data: modifiedGroup });
+      console.log('그룹이 업데이트되었습니다.');
+      console.log(updatedGroup);
+      res.status(200).send({ isSuccess: true, code: 200, data: updatedGroup });
     } else {
+      console.log('그룹을 찾을 수 없습니다.');
       // 업데이트에 실패했을 경우 (상태코드 400과 메시지 반환)
       res.status(400).send({ isSuccess: false, code: 400, error: '업데이트에 실패했습니다.' });
     }
@@ -274,26 +346,75 @@ exports.patchGroupInformation = async (req, res) => {
     res.status(500).send({ isSuccess: false, code: 500, error: err });
   }
 };
-
+//유저 개인이 그룹에서 나올때 (그룹 탈퇴)
 exports.deleteGroup = async (req, res) => {
   try {
+    // 토큰에서 현재 유저 정보 가져오기
+    const userInfo = res.locals.decoded.userInfo;
+
+    if (!userInfo) {
+      return res.status(400).send({
+        isSuccess: false,
+        code: 400,
+        error: '사용자 정보를 찾을 수 없습니다.',
+      });
+    }
+
+    const userId = userInfo.id;
     // 요청 파라미터에서 그룹 ID를 가져옴
     const { groupId } = req.params;
 
     // 그룹 삭제 수행
-    const deleteGroup = await Group.destroy({
-      where: { group_id: groupId },
-    });
+    const deletedGroup = await Group.findById(groupId);
 
-    if (deleteGroup) {
-      // 삭제가 성공했을 경우
-      res.status(204).send({ isSuccess: true, code: 204, msg: '스터디 그룹을 삭제했습니다.' });
+    if (deletedGroup) {
+      // 사용자의 groups 필드에서도 삭제해야 합니다.
+      await User.updateOne({ _id: userId }, { $pull: { groups: groupId } });
+      await Group.updateOne({ _id: groupId }, { $pull: { members: userId } });
+
+      res.status(200).send({ isSuccess: true, code: 204, msg: '해당 스터디 그룹에서 탈퇴했습니다.' });
     } else {
-      // 삭제에 실패했을 경우 (상태코드 400과 에러 메시지 반환)
-      res.status(400).send({ isSuccess: false, code: 400, error: error });
+      // 그룹을 찾지 못한 경우 또는 삭제 실패한 경우
+      console.log('그룹을 찾을 수 없음:', deletedGroup);
+      res.status(400).send({ isSuccess: false, code: 400, error: '그룹 탈퇴에 실패했습니다.' });
     }
   } catch (err) {
+    console.error(err);
     // 서버 오류가 발생한 경우 (상태코드 500과 에러 메시지 반환)
-    res.status(500).send({ isSuccess: false, code: 500, error: err });
+    res.status(500).send({ isSuccess: false, code: 500, error: '서버 오류가 발생했습니다.' });
+  }
+};
+//그룹장이 그룹 삭제
+exports.removeAllMembersFromGroup = async (req, res) => {
+  try {
+    const { groupId } = req.params;
+
+    // 그룹의 members 배열을 비움
+    const updatedGroup = await Group.findByIdAndUpdate(groupId, { $set: { members: [] } });
+
+    if (!updatedGroup) {
+      return res.status(404).send({ isSuccess: false, error: '그룹을 찾을 수 없습니다.' });
+    }
+
+    // 그룹에 속한 모든 멤버 ID 가져오기
+    const groupMembers = updatedGroup.members.map(member => member.toString());
+
+    // 그룹에 대기 중인 모든 사용자 ID 가져오기
+    const pendingGroup = await Group.findByIdAndUpdate(groupId, { $set: { join_requests: [] } });
+    const pendingMembers = pendingGroup.join_requests.map(request => request.user_id.toString());
+
+    // 그룹에서 모든 멤버 삭제
+    await User.updateMany({ _id: { $in: groupMembers } }, { $pull: { groups: groupId } });
+
+    // 대기 중인 그룹 요청을 가진 사용자에서 해당 그룹 삭제
+    await User.updateMany({ _id: { $in: pendingMembers } }, { $pull: { pending_groups: { group: groupId } } });
+
+    // 그룹 스키마의 다큐먼트 삭제
+    await Group.findByIdAndDelete(groupId);
+
+    res.status(200).send({ isSuccess: true, message: '그룹에서 모든 멤버를 삭제했습니다.' });
+  } catch (err) {
+    console.error(err);
+    res.status(500).send({ isSuccess: false, error: '서버 오류가 발생했습니다.' });
   }
 };
